+++ examples/general/src/commonTest/kotlin/com/example/SimpleTestSuite.kt
@@ -12,8 +12,8 @@ val Simple by testSuite {
     }
 }
 
-// Choose a different display name (used on platforms supporting display names).
-val SimpleWithASpecialName by testSuite(displayName = "Suite with a special name") {
+// Choose a name different from the default property FQN
+val SimpleWithASpecialName by testSuite("Suite with a special name") {
     test("string length") {
         assertEquals(8, "Test me!".length)
     }