+++ examples/general/src/commonTest/kotlin/com/example/testLibrary/MochaStyle.kt
@@ -5,8 +5,8 @@ import de.infix.testBalloon.framework.core.TestConfig
 import de.infix.testBalloon.framework.core.TestSuite
 import de.infix.testBalloon.framework.core.TestSuiteScope
 import de.infix.testBalloon.framework.core.testSuite
-import de.infix.testBalloon.framework.shared.TestDisplayName
 import de.infix.testBalloon.framework.shared.TestElementName
+import de.infix.testBalloon.framework.shared.TestElementPropertyFqn
 import de.infix.testBalloon.framework.shared.TestRegistering
 
 fun TestSuiteScope.describe(name: String, testConfig: TestConfig = TestConfig, content: TestSuite.() -> Unit) =
@@ -20,8 +20,8 @@ fun TestSuiteScope.it(
 
 @TestRegistering
 fun describe(
-    @TestElementName name: String = "",
-    @TestDisplayName displayName: String = name,
+    @TestElementName name: String? = null,
     testConfig: TestConfig = TestConfig,
+    @TestElementPropertyFqn propertyFqn: String = "",
     content: TestSuite.() -> Unit
-): Lazy<TestSuite> = testSuite(name = name, displayName = displayName, testConfig = testConfig, content = content)
+): Lazy<TestSuite> = testSuite(name = name, testConfig = testConfig, propertyFqn = propertyFqn, content = content)