+++ integration-test/projectTemplates/element-selection/src/commonTest/kotlin/com/example/ExampleSuites.kt
@@ -23,14 +23,11 @@ val SimpleSuite by testSuite(testConfig = TestConfig.logTestExecution()) {
     }
 }
 
-val DisplayNameSuite by testSuite(
-    displayName = "display d-DisplayNameSuite-d",
-    testConfig = TestConfig.logTestExecution()
-) {
-    test("test 1", displayName = "display test d-1-d") {
+val SpecialNameSuite by testSuite("Should show d-SpecialNameSuite-d", testConfig = TestConfig.logTestExecution()) {
+    test("test 1") {
     }
-    testSuite("suite 1", displayName = "display suite d-1-d") {
-        test("test s1-1", displayName = "display test d-s1-1-d") {
+    testSuite("suite 1") {
+        test("test s1-1") {
         }
     }
 }