+++ testBalloon-compiler-plugin/src/test/kotlin/fakeTestFramework/TestSuite.kt
@@ -2,20 +2,25 @@ package fakeTestFramework
 
 import de.infix.testBalloon.framework.shared.AbstractTestSuite
 import de.infix.testBalloon.framework.shared.TestElementName
+import de.infix.testBalloon.framework.shared.TestElementPropertyFqn
 import de.infix.testBalloon.framework.shared.TestRegistering
 
 @Suppress("unused")
 @TestRegistering
-fun testSuite(@TestElementName name: String = "", content: TestSuite.() -> Unit): Lazy<TestSuite> = lazy {
+fun testSuite(
+    @TestElementName name: String? = null,
+    @TestElementPropertyFqn propertyFqn: String = "",
+    content: TestSuite.() -> Unit
+): Lazy<TestSuite> = lazy {
     TestSuite(
-        name = name,
+        name = name ?: propertyFqn,
+        propertyFqn = propertyFqn,
         content = content
     )
 }
 
-@TestRegistering
-open class TestSuite(@TestElementName name: String = "", content: TestSuite.() -> Unit = {}) :
-    TestElement(name),
+open class TestSuite(name: String = "", propertyFqn: String? = null, content: TestSuite.() -> Unit = {}) :
+    TestElement(name, propertyFqn = propertyFqn),
     AbstractTestSuite {
 
     override var testElementIsEnabled: Boolean = true