+++ testBalloon-framework-core/src/commonTest/kotlin/de/infix/testBalloon/framework/core/TestCompartmentTests.kt
@@ -35,7 +35,7 @@ class TestCompartmentTests {
 
     @Test
     fun twoCompartments() = FrameworkTestUtilities.withTestFramework {
-        val suite1 by testSuite("suite1") {
+        val suite1 by testSuite(propertyFqn = "suite1") {
             val compartment = testElementParent as? TestCompartment
             test("test1") {
                 assertEquals(TestCompartment.Default, compartment)
@@ -50,7 +50,7 @@ class TestCompartmentTests {
         val coroutineName2 = CoroutineName("#2")
         val compartment2 =
             TestCompartment("Compartment 2", testConfig = TestConfig.coroutineContext(coroutineName2))
-        val suite2 by testSuite("suite2", compartment = { compartment2 }) {
+        val suite2 by testSuite(propertyFqn = "suite2", compartment = { compartment2 }) {
             val compartment = testElementParent as? TestCompartment
             test("test1") {
                 assertEquals(compartment2, compartment)
@@ -62,7 +62,7 @@ class TestCompartmentTests {
             }
         }
 
-        val suite3 by testSuite("suite3") {
+        val suite3 by testSuite(propertyFqn = "suite3") {
             val compartment = testElementParent as? TestCompartment
             test("test1") {
                 assertEquals(TestCompartment.Default, compartment)
@@ -74,7 +74,7 @@ class TestCompartmentTests {
             }
         }
 
-        val suite4 by testSuite("suite4", compartment = { compartment2 }) {
+        val suite4 by testSuite(propertyFqn = "suite4", compartment = { compartment2 }) {
             val compartment = testElementParent as? TestCompartment
             test("test1") {
                 assertEquals(compartment, compartment2)
@@ -113,7 +113,7 @@ class TestCompartmentTests {
 
         val concurrentThreadIds = ConcurrentSet<ULong>()
 
-        val suite1 by testSuite("topSuite1", compartment = { TestCompartment.Concurrent }) {
+        val suite1 by testSuite(propertyFqn = "topSuite1", compartment = { TestCompartment.Concurrent }) {
             val outerSuiteThreadId = testPlatform.threadId()
 
             for (suiteNumber in 1..suiteCount) {
@@ -132,7 +132,7 @@ class TestCompartmentTests {
 
         val sequentialThreadIds = ConcurrentSet<ULong>()
 
-        val suite2 by testSuite("topSuite2") {
+        val suite2 by testSuite(propertyFqn = "topSuite2") {
             for (suiteNumber in 1..suiteCount) {
                 testSuite("subSuite$suiteNumber") {
                     for (testNumber in 1..testCount) {
@@ -221,7 +221,9 @@ class TestCompartmentTests {
     fun ui() = FrameworkTestUtilities.withTestFramework {
         val uiThreadIds = ConcurrentSet<ULong>()
 
-        val suite by testSuite("topSuite", compartment = { TestCompartment.MainDispatcher(Dispatchers.Unconfined) }) {
+        val suite by testSuite(propertyFqn = "topSuite", compartment = {
+            TestCompartment.MainDispatcher(Dispatchers.Unconfined)
+        }) {
             for (suiteNumber in 1..3) {
                 testSuite("subSuite$suiteNumber") {
                     test("test1") {