+++ testBalloon-framework-core/src/commonTest/kotlin/de/infix/testBalloon/framework/core/TestConfigTests.kt
@@ -14,7 +14,7 @@ import kotlin.test.assertNull
 class TestConfigTests {
     @Test
     fun coroutineContext() = FrameworkTestUtilities.withTestFramework {
-        val testSuite by testSuite("testSuite") {}
+        val testSuite by testSuite(propertyFqn = "testSuite") {}
 
         TestConfig.executeWrapped(testSuite) {
             assertNull(currentCoroutineContext()[CoroutineName.Key])
@@ -28,7 +28,7 @@ class TestConfigTests {
 
     @Test
     fun invocation() = FrameworkTestUtilities.withTestFramework {
-        val testSuite by testSuite("testSuite") {}
+        val testSuite by testSuite(propertyFqn = "testSuite") {}
 
         TestConfig.executeWrapped(testSuite) {
             assertEquals(TestConfig.Invocation.Sequential, TestConfig.Invocation.current())
@@ -41,7 +41,7 @@ class TestConfigTests {
 
     @Test
     fun concurrencyAndTestScope() = FrameworkTestUtilities.withTestFramework {
-        val testSuite by testSuite("testSuite") {}
+        val testSuite by testSuite(propertyFqn = "testSuite") {}
 
         TestSession.DefaultConfiguration.executeWrapped(testSuite) {
             assertEquals(TestConfig.Invocation.Sequential, TestConfig.Invocation.current())
@@ -121,7 +121,7 @@ class TestConfigTests {
 
     private fun testPermits(testConfig: TestConfig, vararg expectedPermits: TestConfig.Permit) =
         FrameworkTestUtilities.withTestFramework {
-            val testSuite by testSuite("testSuite") {}
+            val testSuite by testSuite(propertyFqn = "testSuite") {}
 
             testConfig.parameterize(testSuite)
 
@@ -192,7 +192,7 @@ class TestConfigTests {
 
     private fun testKeyedParameters(testConfig: TestConfig, vararg expectedParameters: TestElement.KeyedParameter) =
         FrameworkTestUtilities.withTestFramework {
-            val testSuite by testSuite("testSuite") {}
+            val testSuite by testSuite(propertyFqn = "testSuite") {}
 
             testConfig.parameterize(testSuite)
 
@@ -203,7 +203,7 @@ class TestConfigTests {
 
     @Test
     fun testScope() = FrameworkTestUtilities.withTestFramework {
-        val testSuite by testSuite("testSuite") {}
+        val testSuite by testSuite(propertyFqn = "testSuite") {}
 
         TestConfig.executeWrapped(testSuite) {
             assertNull(TestScopeContext.current())
@@ -220,7 +220,7 @@ class TestConfigTests {
 
     @Test
     fun inheritance() = FrameworkTestUtilities.withTestFramework {
-        val testSuite by testSuite("testSuite") {}
+        val testSuite by testSuite(propertyFqn = "testSuite") {}
 
         val coroutineNameElement = CoroutineName("TEST-CC")
         TestConfig.coroutineContext(coroutineNameElement).executeWrapped(testSuite) {
@@ -235,7 +235,7 @@ class TestConfigTests {
 
     @Test
     fun chaining() = FrameworkTestUtilities.withTestFramework {
-        val testSuite by testSuite("testSuite") {}
+        val testSuite by testSuite(propertyFqn = "testSuite") {}
 
         val coroutineNameElement = CoroutineName("TEST-CC")
         assertNull(currentCoroutineContext()[CoroutineName.Key])