+++ testBalloon-framework-shared/src/commonMain/kotlin/de/infix/testBalloon/framework/shared/Annotations.kt
@@ -10,14 +10,9 @@ public annotation class TestRegistering
 /**
  * Designates a test-registering function's String parameter as representing the test element's name.
  *
- * For a top-level [TestRegistering] function, this implies:
- * - A `@TestElementName` parameter must have a default value (typically, "").
- * - If the call site of the function does not supply an actual value for that `@TestElementName` parameter,
- *   the compiler will insert the caller's fully qualified property name.
- *
- * In addition, the IDE plugin uses this annotation to detect the element name position in a parameter list.
- * The optional [prefix] and [postfix] parameters make the IDE plugin aware of modifications to the element name,
- * so that a valid test element path can be produced by static analysis. Example for a Behavior/Gherkin-style
+ * The IDE plugin uses this annotation to detect the function parameter containing the element name.
+ * The optional [prefix] and [postfix] annotation parameters make the IDE plugin aware of modifications to the element
+ * name, so that a valid test element path can be produced by static analysis. Example for a Behavior/Gherkin-style
  * `Scenario` function creating a test suite:
  * ```
  * @TestRegistering
@@ -39,15 +34,15 @@ public annotation class TestRegistering
 public annotation class TestElementName(val prefix: String = "", val postfix: String = "")
 
 /**
- * Makes a String parameter receive the simple name of its call-site top-level property.
+ * Designates the String parameter of a top-level test suite function which receives the FQN of the property.
  *
- * For prerequisites and mechanism, see [TestElementName].
- *
- * The display name generated is the simple name of the caller's class or property.
+ * A top-level [TestRegistering] test suite function must have exactly one `String?` parameter annotated with
+ * `@TestElementPropertyFqn` with a default value of `""`. At the call site of such function, the compiler plugin will
+ * insert the fully qualified name of the corresponding top-level property.
  */
 @Target(AnnotationTarget.VALUE_PARAMETER)
 @Retention(AnnotationRetention.BINARY)
-public annotation class TestDisplayName
+public annotation class TestElementPropertyFqn
 
 /**
  * Makes the annotated class part of the test element DSL. See the Kotlin docs on [DslMarker] for details.